Creates printNote, loadNotes, saveNotes, addNote, and deleteNote.
"printNote" takes a note object as argument and prints to terminal in readable format the title and body. "loadNotes" returns the JSON parsed contents of the file path 'notes-data.json'; when the file is null an empty list is returned. "saveNotes" stores a string version of the notes list it receives as argument. "addNote" takes as argument a title and body; it checks if the title is already stored in data if not it creates a new note adds it to the existing list of notes and calls "saveNotes".